refactor : UserDto,ArticleDto를 도메인으로 사용
