--- src/main/java/kr/codesqaud/cafe/dto/UserDto.java
@@ -1,60 +1,0 @@
-package kr.codesqaud.cafe.dto;
-
-import kr.codesqaud.cafe.domain.member.UserLog;
-
-public class UserDto {
-
-    private int index;
-    private String userId;
-    private String password;
-    private String name;
-    private String email;
-
-    public UserDto( String userId, String password, String name, String email) {
-        this.userId = userId;
-        this.password = password;
-        this.name = name;
-        this.email = email;
-        this.index = UserLog.userLogList.size();
-    }
-
-    public int getIndex() {
-        return index;
-    }
-
-    public void setIndex(int index) {
-        this.index = index;
-    }
-
-    public String getUserId() {
-        return userId;
-    }
-
-    public void setUserId(String userId) {
-        this.userId = userId;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-}
